💚 fix the RtD doxygen run (#912)
## Description

This PR fixes the doxygen run that is performed on RtD.
This would previously fail silently because it could not create the
respective directory.
This also adds a release note to prepare for the imminent patch release.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>