📝 Correct and add project metadata (#911)
## Description

This PR fixes a couple of links in the `[project.urls]` table that is
part of the Python package metadata.
This ensures that the respective links appear as verified on PyPI and
was an oversight after moving the repository.
This PR also adds the new Changelog to the list of URLs.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>