📝 Add JOSS journal reference and citation information (#913)
## Description

This PR adds the journal reference from the just-published article in
the Journal of Open Source Software 🎉
It also adds the respective citation information and touches up the
changelog.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>