🐛 Fix sxdg support in qiskit import (#930)
## Description

This PR fixes an oversight in the Qiskit import functionality, where
`sxdg` would not be supported.
The `SX` gate family was missing in the corresponding test as well.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>