♻️ [MLIR] Reduce Redundancy in Operation Definition (#933)
## Description

In the previous infrastructure of MLIR, we defined all Standard
Operations in every dialect separately. This led to duplicate code. This
PR mitigates this issue by outsourcing the definitions of the Standard
Operations (that are identical on a character level). It builds up
infrastructure such that they are configured for a corresponding dialect
and can be included in the tablegen file with the definitions of the
operations for each dialect.

Additionally, this PR renames all occurrences of `Qureg`or `qureg` to
`Qreg` and `qreg`, respectively.

## Checklist:

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: Yannick Stade <100073938+ystade@users.noreply.github.com>
Signed-off-by: Lukas Burgholzer <burgholzer@me.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>