🩹 Fix MLIR workflow (#937)
## Description

#936 broke the MLIR workflow. This went unnoticed because the MLIR
workflow would not trigger on workflow-only changes.
This PR fixes the oversight and enables the CI to run on workflow-only
changes, which should be demonstrated by this PR.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>