:construction_worker: add windows CI for MLIR (#952)
## Description

This PR adds MLIR CI tests for windows runners.
This is achieved by building MLIR from source so the first execution
should take a bit longer.
Afterwards, the cache will be used to store the built versions of MLIR.

This PR contributes to #925

**Note:** To install a specific version, the runner clones the tag
`llvmorg-{version}.1.0`. It seems that the first release of a new major
`llvm-project` version is typically called `{version}.1.0` so this
should keep working, but there is no guarantee.

## Checklist:

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>