⚗️ Testing functionality of MQTOpt value dialect (#948)
## Description

As described in #899, we want to improve the testing setup and coverage
across the MLIR part of the Core library. The purpose of this PR is to
have a structured approach to testing all functionalities of the MQTOpt
dialect. In other words, we want to make sure that each operation
behaves as expected.

This PR is only a contribution to #899, not a fix.

## Checklist:

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes or removals.
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.

---------

Signed-off-by: Lukas Burgholzer <burgholzer@me.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>