:rotating_light: Use -march instead of -mcpu for GCC on x86
Using -mcpu has been necessary for ARM-based macOS systems. However,
this option is deprecated on x86 platforms and leads to a warning for
every source file.
This adds an additional platform check and uses -march if the system is
x86 based.

Since the value of CMAKE_SYSTEM_PROCESSOR is not standardized, it might
become necessary to add additional values to the regex.