♻️ Specify Install Components for Wheel (#1021)
## Description

This PR does not immediately solve an issue but contributes to avoiding
an issue that arose in #996. In particular, it allows to select the
install components explicitly and, hence, avoid installing components
that were added by dependencies where install instructions cannot be
deactivated, e.g., RE2.

Also in general, I think, it is a good idea to set the install
components for the wheel explicitly.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [ ] I have added appropriate tests that cover the new/changed
functionality.
- [ ] I have updated the documentation to reflect these changes.
- [ ] I have added entries to the changelog for any noteworthy
additions, changes, fixes or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: Yannick Stade <100073938+ystade@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>