🔧 Update CMake configuration for MLIR subproject (#1043)
## Description

This PR takes some of the lessons learned from working on #881 and
adjusts the CMake configuration for the MLIR subproject.

In particular, it fixes the header file sets to also include the header
generated in the build folder.
It also fixes some of the CMake variable used in the `SetupMLIR.cmake`
script so that the script still works, when MQT Core is used as a
submodule.
Lastly, this PR turns the `CMAKE_CXX_STANDARD` into a cache variable,
which seems to be the recommended best practice because it allows
overriding the standard from a top-level project.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>