⬆️ raise `pybind11` lower bound to `v3.0.0` (#1050)
<!--- This file has been generated from an external template. Please do
not modify it directly. -->
<!--- Changes should be contributed to
https://github.com/munich-quantum-toolkit/templates. -->

## Description

This PR updates the `pybind11` lower bound to v3.0.0.
At the moment, we do not make use of any of the latest major release's
features.
However, this is clearly the plan. Especially the new native enum
support as well as the better compatibility between extensions compiled
with different compiler versions.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

Signed-off-by: burgholzer <burgholzer@me.com>