✨ MLIR - Add `allocQubit` operations to `MQTOpt` and `MQTRef` (#1139)
## Description

Adds new operations for the allocation (and de-allocation) of qubits
without the use of qubit registers to the MLIR dialects.

This allows for a more streamlined and explicit pipeline to access
qubits.
In a future update, the corresponding `QubitRegister` types and
operations might be removed to focus *only* on individual qubits. In
that case, the `alloc` operations might be refined by allowing multiple
qubits to be allocated at once, as in
```
%q0, %q1, %q2 = mqtopt.allocQubit() <{count_attr = 3 : i64}> : () -> !mqtopt.Qubit, !mqtopt.Qubit, !mqtopt.Qubit
```

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>