💚🏁 Fix path management for Windows MLIR tests (#1141)
<!--- This file has been generated from an external template. Please do
not modify it directly. -->
<!--- Changes should be contributed to
https://github.com/munich-quantum-toolkit/templates. -->

## Description

This PR aims to fix the MLIR CI that was unintentionally broken in #1139
due to our current lack of branch protection for the MLIR part of the
test suite.
The necessary changes to enable that will first be added upstream in the
workflows repository.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>