✨ Collect clifford blocks (#885)
## Description

Added the ability to separate the collection of Clifford and non
Clifford Blocks by adding a new function parameter. For that a
isCliffordCheck was added as well. The isClifford check is very rough
but includes all common Clifford gates. However, it does not include
rotation gates that with the right angle could be considered Clifford.
It does not include custom gates that could be Clifford as well.

The separation is done by finalizing a block when we iterate over a
non-Clifford Block. When we want to collect Clifford and non Clifford
gates separately then the function handles non-Clifford gates as "cannot
be processed" and the canProcess variable is set to false. This ensures
that we still have large CliffordBlocks but we have only single non
CliffordBlocks. Essentially separating the non-CliffordBlocks from all
Clifford and other non-CliffordBlocks.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: Jannik Pflieger <56534984+jannikpflieger@users.noreply.github.com>
Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: burgholzer <burgholzer@me.com>