🍎🔥 Stop testing on `x86` macOS and shipping wheels for Intel-based Macs (#1165)
## Description

This PR drops running our tests and deployment workflows on `macos-13`
runners.
Hence, this marks the end of us shipping macOS x86 wheels with MQT Core.
While I do not expect any immediate problems and incompatibilities, this
means that people still on those systems will either have to stick with
an old version of MQT Core or will have to build the new version from
source.

I'll let this sit in draft for a while until it becomes relevant.
If nothing bigger comes up, this will be part of the next minor release.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>