🎨 Replacement of 'friend' in FoMaC with Token Pattern (#1186)
## Description

In #1150 it was mentioned that the acces to private constructors in the
FoMaC library could be implemented differently, see comment
https://github.com/munich-quantum-toolkit/core/pull/1150#discussion_r2317454567
.

This PR implements a proposal to replace the 'friend' declaration with
the Token pattern and nests the classes appropriately.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>