✨ Add MQTRef-QIR conversions (#1091)
<!--- This file has been generated from an external template. Please do
not modify it directly. -->
<!--- Changes should be contributed to
https://github.com/munich-quantum-toolkit/templates. -->

## Description

This PR introduces conversion passes between the `mqtdyn` dialect and
the `LLVM IR` that adheres to the QIR specifications for dynamic qubit
allocations. Currently it mainly focuses on the conversion within MLIR
between the `mqtdyn` dialect and the `llvmlir` dialect and vice versa.
Once the IR consists only of llvm dialect operations it can be convert
to LLVM IR with the built-in passes from the `mlir-translate` module and
back.


## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Signed-off-by: li-mingbao <74404929+li-mingbao@users.noreply.github.com>
Co-authored-by: Li-ming Bao <ge93vof@tum.de>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Yannick Stade <100073938+ystade@users.noreply.github.com>
Co-authored-by: burgholzer <burgholzer@me.com>